+++ client/src/app/layouts/DefaultLayout.jsx
@@ -3,6 +3,7 @@ import { AppSidebar } from "@/components/app-sidebar";
 import { SiteHeader } from "@/components/site-header";
 
 import { SidebarProvider, SidebarInset } from "@/ui/sidebar";
+import { ScrollArea } from "@/shared/components/ui/scroll-area";
 
 export default function DefaultLayout() {
   return (
@@ -15,9 +16,13 @@ export default function DefaultLayout() {
       <AppSidebar variant="inset" />
       <SidebarInset>
         <SiteHeader />
-        <main className="flex-1 flex flex-col">
-          <Outlet />
-        </main>
+        <ScrollArea
+          className="flex-1 h-[calc(100vh-var(--header-height))]"
+          viewportClassName="h-full">
+          <main className="flex min-h-full flex-col">
+            <Outlet />
+          </main>
+        </ScrollArea>
       </SidebarInset>
     </SidebarProvider>
   );